Added LookupTable class Added SmartEnums Added Unit tests for SmartEnums Fixed GE_LOG, created an int32 from __LINE__, because in Windows this value might pass as a timestamp

+++ sdk/geUtilities/include/geDebug.h
@@ -185,7 +185,7 @@ bool LogCategory##name::s_registered = Log::_registerCategory(LogCategory##name:
                           String("\n\t\t in ") +                              \
                           __PRETTY_FUNCTION__ +                               \
                           " [" + __FILE__ + ":" +                             \
-                          toString(__LINE__) + "]\n",                         \
+                          toString(int32(__LINE__)) + "]\n",                  \
                   LogVerbosity::verbosity, LogCategory##category::_id);       \
   }} while (0)
